+++ src/main/java/me/clip/placeholderapi/commands/command/ecloud/PlaceholdersCommand.java
@@ -4,8 +4,8 @@ import me.clip.placeholderapi.PlaceholderAPI;
 import me.clip.placeholderapi.PlaceholderAPIPlugin;
 import me.clip.placeholderapi.commands.Command;
 import me.clip.placeholderapi.expansion.cloud.CloudExpansion;
+import me.clip.placeholderapi.util.JSONMessage;
 import me.clip.placeholderapi.util.Msg;
-import me.rayzr522.jsonmessage.JSONMessage;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
